Refactor controllers, entities, DTOs and services
Add new REST controllers (AdopterController, PetController) and standardize endpoint names and response handling across controllers; ReviewController now uses ReviewDTO and maps entities to DTOs. Update DTOs (AdoptionDTO, AdoptionDetailDTO, ReviewDetailDTO, ShelterEventDTO) and add a toEntity helper for AdoptionDTO. Rework JPA entities: adjust types, initialize collections, change Lombok annotations, add JsonIgnore/JsonIgnoreProperties to prevent serialization/lazy-loading issues, and refine relationship mappings (e.g. eventCode -> Long, eager/lazy fetch adjustments). Services updated with transactional annotations, stronger validation and business rules (AdopterService and AdoptionService improved checks and messages, PetService simplified and made transactional). Minor repository/import cleanups and controller response consistency (use ResponseEntity.ok / noContent). These changes address serialization errors, NPEs, and tighten business logic and validation.

+++ src/main/java/co/edu/udistrital/mdp/pets/dto/ShelterEventDTO.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 public class ShelterEventDTO {
     
     private Long id;
-    private Integer eventCode;
+    private Long eventCode;
     private String eventType;
     private String title;
     private String description;
